# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RoomPlayer {
  id: string;
  roomId: number;
  name: string;
  gender: string;
  skill: number;
  speed: number;
  attending: boolean;
  createdAt: string;
}

export interface Room {
  id: number;
  code: string;
  createdAt: string;
}

export interface RoomState {
  room: Room;
  players: RoomPlayer[];
}

export interface PlayerInput {
  id: string;
  /** @minLength 1 */
  name: string;
  gender: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  skill: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  speed: number;
}

export interface PlayerUpdate {
  /** @minLength 1 */
  name?: string;
  gender?: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  skill?: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  speed?: number;
  attending?: boolean;
}

export interface AttendanceBulk {
  attending: boolean;
}