# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: rooms
    description: Collaborative room operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /rooms:
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new collaborative room
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomState"

  /rooms/{code}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room with all players
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Room state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomState"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{code}/players:
    post:
      operationId: addPlayer
      tags: [rooms]
      summary: Add a player to the room
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Player added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomPlayer"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{code}/players/{playerId}:
    patch:
      operationId: updatePlayer
      tags: [rooms]
      summary: Update a player (skill, speed, attending, etc.)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Player updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomPlayer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removePlayer
      tags: [rooms]
      summary: Remove a player from the room
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Player removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /rooms/{code}/attendance:
    put:
      operationId: setAllAttendance
      tags: [rooms]
      summary: Set all players attending or not (Select All / Deselect All)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceBulk"
      responses:
        "200":
          description: Attendance updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomState"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    RoomPlayer:
      type: object
      required: [id, roomId, name, gender, skill, speed, attending, createdAt]
      properties:
        id:
          type: string
        roomId:
          type: integer
        name:
          type: string
        gender:
          type: string
        skill:
          type: integer
        speed:
          type: integer
        attending:
          type: boolean
        createdAt:
          type: string
          format: date-time

    Room:
      type: object
      required: [id, code, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        createdAt:
          type: string
          format: date-time

    RoomState:
      type: object
      required: [room, players]
      properties:
        room:
          $ref: "#/components/schemas/Room"
        players:
          type: array
          items:
            $ref: "#/components/schemas/RoomPlayer"

    PlayerInput:
      type: object
      required: [id, name, gender, skill, speed]
      properties:
        id:
          type: string
        name:
          type: string
          minLength: 1
        gender:
          type: string
        skill:
          type: integer
          minimum: 0
          maximum: 10
        speed:
          type: integer
          minimum: 1
          maximum: 3

    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        gender:
          type: string
        skill:
          type: integer
          minimum: 0
          maximum: 10
        speed:
          type: integer
          minimum: 1
          maximum: 3
        attending:
          type: boolean

    AttendanceBulk:
      type: object
      required: [attending]
      properties:
        attending:
          type: boolean