# lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerInput {
  id: string;
  /** @minLength 1 */
  name: string;
  gender: string;
  /**
     * @minimum 0
     * @maximum 10
     */
  skill: number;
  /**
     * @minimum 1
     * @maximum 3
     */
  speed: number;
}