# lib/api-zod/src/generated/types/roomPlayer.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomPlayer {
  id: string;
  roomId: number;
  name: string;
  gender: string;
  skill: number;
  speed: number;
  attending: boolean;
  createdAt: Date;
}