# lib/api-zod/src/generated/types/roomState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Room } from './room';
import type { RoomPlayer } from './roomPlayer';

export interface RoomState {
  room: Room;
  players: RoomPlayer[];
}